Missed a spot

+++ vice/src/lib/libusbsiddrv/USBSIDInterface.cpp
@@ -117,7 +117,7 @@ extern "C"
     if( p == NULL ) return -1;
     return ((USBSID_Class*)p)->us_Found;
   }
-  void writesingle_USBSID(USBSIDitf p, unsigned char *buff, size_t len){
+  void writesingle_USBSID(USBSIDitf p, unsigned char *buff, int len){
     if( p == NULL ) return;
     return ((USBSID_Class*)p)->USBSID_SingleWrite(buff, len);
   };
